requirements.txt
numpy==1.20.0
pandas==1.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
